+++ Sources/AppwriteEnums/AuthenticationFactor.swift
@@ -1,6 +1,6 @@
 import Foundation
 
-public enum AuthenticationFactor: String, CustomStringConvertible {
+public enum AuthenticationFactor: String, Codable, CustomStringConvertible {
     case email = "email"
     case phone = "phone"
     case totp = "totp"