+++ Sources/AppwriteModels/Presence.swift
@@ -1,107 +1,116 @@
 import Foundation
 import JSONCodable
 
-/// Row
-open class Row<T : Codable>: Codable {
+/// Presence
+open class Presence<T : Codable>: Codable {
 
     enum CodingKeys: String, CodingKey {
         case id = "$id"
-        case sequence = "$sequence"
-        case tableId = "$tableId"
-        case databaseId = "$databaseId"
         case createdAt = "$createdAt"
         case updatedAt = "$updatedAt"
         case permissions = "$permissions"
-        case data = "data"
+        case userId = "userId"
+        case status = "status"
+        case source = "source"
+        case expiresAt = "expiresAt"
+        case metadata = "metadata"
     }
 
-    /// Row ID.
+    /// Presence ID.
     public let id: String
-    /// Row sequence ID.
-    public let sequence: String
-    /// Table ID.
-    public let tableId: String
-    /// Database ID.
-    public let databaseId: String
-    /// Row creation date in ISO 8601 format.
+    /// Presence creation date in ISO 8601 format.
     public let createdAt: String
-    /// Row update date in ISO 8601 format.
+    /// Presence update date in ISO 8601 format.
     public let updatedAt: String
-    /// Row permissions. [Learn more about permissions](https://appwrite.io/docs/permissions).
+    /// Presence permissions. [Learn more about permissions](https://appwrite.io/docs/permissions).
     public let permissions: [String]
+    /// User ID.
+    public let userId: String
+    /// Presence status.
+    public let status: String?
+    /// Presence source.
+    public let source: String
+    /// Presence expiry date in ISO 8601 format.
+    public let expiresAt: String?
     /// Additional properties
-    public let data: T
+    public let metadata: T
 
     init(
         id: String,
-        sequence: String,
-        tableId: String,
-        databaseId: String,
         createdAt: String,
         updatedAt: String,
         permissions: [String],
-        data: T
+        userId: String,
+        status: String?,
+        source: String,
+        expiresAt: String?,
+        metadata: T
     ) {
         self.id = id
-        self.sequence = sequence
-        self.tableId = tableId
-        self.databaseId = databaseId
         self.createdAt = createdAt
         self.updatedAt = updatedAt
         self.permissions = permissions
-        self.data = data
+        self.userId = userId
+        self.status = status
+        self.source = source
+        self.expiresAt = expiresAt
+        self.metadata = metadata
     }
 
     public required init(from decoder: Decoder) throws {
         let container = try decoder.container(keyedBy: CodingKeys.self)
 
         self.id = try container.decode(String.self, forKey: .id)
-        self.sequence = try container.decode(String.self, forKey: .sequence)
-        self.tableId = try container.decode(String.self, forKey: .tableId)
-        self.databaseId = try container.decode(String.self, forKey: .databaseId)
         self.createdAt = try container.decode(String.self, forKey: .createdAt)
         self.updatedAt = try container.decode(String.self, forKey: .updatedAt)
         self.permissions = try container.decode([String].self, forKey: .permissions)
-        self.data = try container.decode(T.self, forKey: .data)
+        self.userId = try container.decode(String.self, forKey: .userId)
+        self.status = try container.decodeIfPresent(String.self, forKey: .status)
+        self.source = try container.decode(String.self, forKey: .source)
+        self.expiresAt = try container.decodeIfPresent(String.self, forKey: .expiresAt)
+        self.metadata = try container.decode(T.self, forKey: .metadata)
     }
 
     public func encode(to encoder: Encoder) throws {
         var container = encoder.container(keyedBy: CodingKeys.self)
 
         try container.encode(id, forKey: .id)
-        try container.encode(sequence, forKey: .sequence)
-        try container.encode(tableId, forKey: .tableId)
-        try container.encode(databaseId, forKey: .databaseId)
         try container.encode(createdAt, forKey: .createdAt)
         try container.encode(updatedAt, forKey: .updatedAt)
         try container.encode(permissions, forKey: .permissions)
-        try container.encode(data, forKey: .data)
+        try container.encode(userId, forKey: .userId)
+        try container.encodeIfPresent(status, forKey: .status)
+        try container.encode(source, forKey: .source)
+        try container.encodeIfPresent(expiresAt, forKey: .expiresAt)
+        try container.encode(metadata, forKey: .metadata)
     }
 
     public func toMap() -> [String: Any] {
         return [
             "$id": id as Any,
-            "$sequence": sequence as Any,
-            "$tableId": tableId as Any,
-            "$databaseId": databaseId as Any,
             "$createdAt": createdAt as Any,
             "$updatedAt": updatedAt as Any,
             "$permissions": permissions as Any,
-            "data": (try! JSONSerialization.jsonObject(with: JSONEncoder().encode(data))) as? [String: Any] ?? [:]
+            "userId": userId as Any,
+            "status": status as Any,
+            "source": source as Any,
+            "expiresAt": expiresAt as Any,
+            "metadata": (try! JSONSerialization.jsonObject(with: JSONEncoder().encode(metadata))) as? [String: Any] ?? [:]
         ]
     }
 
-    public static func from(map: [String: Any] ) -> Row {
-        return Row(
+    public static func from(map: [String: Any] ) -> Presence {
+        return Presence(
             id: map["$id"] as! String,
-            sequence: map["$sequence"] as! String,
-            tableId: map["$tableId"] as! String,
-            databaseId: map["$databaseId"] as! String,
             createdAt: map["$createdAt"] as! String,
             updatedAt: map["$updatedAt"] as! String,
             permissions: map["$permissions"] as! [String],
-            data: try! JSONDecoder().decode(T.self, from: {
-                let raw = map["data"]
+            userId: map["userId"] as! String,
+            status: map["status"] as? String,
+            source: map["source"] as! String,
+            expiresAt: map["expiresAt"] as? String,
+            metadata: try! JSONDecoder().decode(T.self, from: {
+                let raw = map["metadata"]
                 if let dict = raw as? [String: Any] {
                     return try! JSONSerialization.data(withJSONObject: dict, options: [])
                 } else if let raw = raw, JSONSerialization.isValidJSONObject(raw) {